# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all tasks
 */
export const ListTasksQueryParams = zod.object({
  status: zod.enum(["all", "pending", "completed"]).optional(),
});

export const ListTasksResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().nullish(),
  completed: zod.boolean(),
  priority: zod.enum(["low", "medium", "high"]),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListTasksResponse = zod.array(ListTasksResponseItem);

/**
 * @summary Create a task
 */
export const CreateTaskBody = zod.object({
  title: zod.string(),
  description: zod.string().nullish(),
  priority: zod.enum(["low", "medium", "high"]),
});

/**
 * @summary Get a task by ID
 */
export const GetTaskParams = zod.object({
  id: zod.coerce.number(),
});

export const GetTaskResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().nullish(),
  completed: zod.boolean(),
  priority: zod.enum(["low", "medium", "high"]),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update a task
 */
export const UpdateTaskParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateTaskBody = zod.object({
  title: zod.string().optional(),
  description: zod.string().nullish(),
  completed: zod.boolean().optional(),
  priority: zod.enum(["low", "medium", "high"]).optional(),
});

export const UpdateTaskResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().nullish(),
  completed: zod.boolean(),
  priority: zod.enum(["low", "medium", "high"]),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a task
 */
export const DeleteTaskParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get task statistics summary
 */
export const GetTasksSummaryResponse = zod.object({
  total: zod.number(),
  completed: zod.number(),
  pending: zod.number(),
  highPriority: zod.number(),
  completionRate: zod.number(),
});